README.md: Add some more notes to the mock data setup.
src/app/app.module.ts:
src/app/client-list/client.service.ts:
Make a live connection to a local MongoDB instance through a REST API call, authenticated using "admin/changeit", the default restheart credentials. Note that it's using an actual host name, because I launched MongoDB and restheart as microservices in a docker-compose back-end, with port 8080 exposed, but it seemed not to be happy to accept connections over localhost.
Never mind; the next step is to create an image containing this Angular web app, and include it in the docker-compose setup.