src/app/client/client.ts: Set the Client up with a constructor so that an instance can be constructed and POSTed to the API to add a new client.
src/app/client/client.service.ts:
Create a Client, with "today" as the default registration date, when adding a new client. The idea is that the registration date is not needed in the GUI: it defaults to the day he GUI is submitted, but we can always expose it later on if clients have to be added retrospectively.